# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  name: string;
  email: string;
  phone: string;
  shipFirstName: string;
  shipLastName: string;
  shipAddress1: string;
  shipAddress2: string;
  shipCity: string;
  shipState: string;
  shipZip: string;
  shipCountry: string;
  billSameAsShip: boolean;
  billFirstName: string;
  billLastName: string;
  billAddress1: string;
  billAddress2: string;
  billCity: string;
  billState: string;
  billZip: string;
  billCountry: string;
  addressJigEnabled: boolean;
  costcoMembershipId: string;
  imapHost: string;
  imapPort: string;
  imapUser: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProfileBody {
  name: string;
  email: string;
  phone?: string;
  shipFirstName?: string;
  shipLastName?: string;
  shipAddress1?: string;
  shipAddress2?: string;
  shipCity?: string;
  shipState?: string;
  shipZip?: string;
  shipCountry?: string;
  billSameAsShip?: boolean;
  billFirstName?: string;
  billLastName?: string;
  billAddress1?: string;
  billAddress2?: string;
  billCity?: string;
  billState?: string;
  billZip?: string;
  billCountry?: string;
  addressJigEnabled?: boolean;
  costcoMembershipId?: string;
  imapHost?: string;
  imapPort?: string;
  imapUser?: string;
  imapPassword?: string;
}

export interface UpdateProfileBody {
  name?: string;
  email?: string;
  phone?: string;
  shipFirstName?: string;
  shipLastName?: string;
  shipAddress1?: string;
  shipAddress2?: string;
  shipCity?: string;
  shipState?: string;
  shipZip?: string;
  shipCountry?: string;
  billSameAsShip?: boolean;
  billFirstName?: string;
  billLastName?: string;
  billAddress1?: string;
  billAddress2?: string;
  billCity?: string;
  billState?: string;
  billZip?: string;
  billCountry?: string;
  addressJigEnabled?: boolean;
  costcoMembershipId?: string;
  imapHost?: string;
  imapPort?: string;
  imapUser?: string;
  imapPassword?: string;
}

export interface CreditCard {
  id: number;
  profileId: number;
  cardNickname: string;
  cardholderName: string;
  lastFour: string;
  cardType: string;
  expiryMonth: string;
  expiryYear: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCreditCardBody {
  profileId: number;
  cardNickname?: string;
  cardholderName: string;
  cardNumber: string;
  cvv: string;
  expiryMonth: string;
  expiryYear: string;
}

export interface UpdateCreditCardBody {
  cardNickname?: string;
  cardholderName?: string;
  cardNumber?: string;
  cvv?: string;
  expiryMonth?: string;
  expiryYear?: string;
}

export interface Proxy {
  id: number;
  label: string;
  host: string;
  port: string;
  username: string;
  isActive: boolean;
  lastTestStatus: string;
  lastTestLatency: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProxyBody {
  label?: string;
  host: string;
  port: string;
  username?: string;
  password?: string;
  isActive?: boolean;
}

export interface UpdateProxyBody {
  label?: string;
  host?: string;
  port?: string;
  username?: string;
  password?: string;
  isActive?: boolean;
  lastTestStatus?: string;
  lastTestLatency?: string;
}

export interface ProxyTestResult {
  success: boolean;
  latency: string;
  ip: string;
  message: string;
}

export interface TaskGroup {
  id: number;
  name: string;
  retailer: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTaskGroupBody {
  name: string;
  retailer: string;
}

export interface UpdateTaskGroupBody {
  name?: string;
  retailer?: string;
}

export interface Task {
  id: number;
  /** @nullable */
  groupId?: number | null;
  /** @nullable */
  profileId?: number | null;
  /** @nullable */
  proxyId?: number | null;
  retailer: string;
  productUrl: string;
  productKeywords: string;
  size: string;
  quantity: number;
  monitorDelay: number;
  retryCount: number;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTaskBody {
  groupId?: number;
  profileId?: number;
  proxyId?: number;
  retailer: string;
  productUrl?: string;
  productKeywords?: string;
  size?: string;
  quantity?: number;
  monitorDelay?: number;
  retryCount?: number;
}

export interface UpdateTaskBody {
  groupId?: number;
  profileId?: number;
  proxyId?: number;
  retailer?: string;
  productUrl?: string;
  productKeywords?: string;
  size?: string;
  quantity?: number;
  monitorDelay?: number;
  retryCount?: number;
  status?: string;
}

export interface BulkActionResult {
  affected: number;
  message: string;
}

export interface CheckoutResult {
  id: number;
  /** @nullable */
  taskId?: number | null;
  success: boolean;
  productName: string;
  productImage: string;
  /** @nullable */
  price?: string | null;
  retailer: string;
  orderNumber: string;
  errorMessage: string;
  /** @nullable */
  profileId?: number | null;
  createdAt: string;
}

export interface CreateCheckoutResultBody {
  taskId?: number;
  success: boolean;
  productName?: string;
  productImage?: string;
  price?: string;
  retailer: string;
  orderNumber?: string;
  errorMessage?: string;
  profileId?: number;
}

export interface AnalyticsSummary {
  totalCheckouts: number;
  totalFailures: number;
  totalSpent: string;
  totalSaved: string;
  successRate: number;
  activeTasks: number;
  totalProfiles: number;
  totalProxies: number;
}

export interface CheckoutTimeSeriesPoint {
  date: string;
  checkouts: number;
  failures: number;
}

export type ListCreditCardsParams = {
  profileId?: number;
};

export type ListTasksParams = {
  groupId?: number;
};

export type ListCheckoutResultsParams = {
  limit?: number;
  success?: boolean;
};

export type GetCheckoutsOverTimeParams = {
  period?: GetCheckoutsOverTimePeriod;
};

export type GetCheckoutsOverTimePeriod =
  (typeof GetCheckoutsOverTimePeriod)[keyof typeof GetCheckoutsOverTimePeriod];

export const GetCheckoutsOverTimePeriod = {
  day: "day",
  week: "week",
  month: "month",
  year: "year",
} as const;