# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: groupId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tasks/{id}/start:
    post:
      operationId: startTask
      tags: [tasks]
      summary: Start a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/stop:
    post:
      operationId: stopTask
      tags: [tasks]
      summary: Stop a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: ["integer", "null"]
        profileId:
          type: ["integer", "null"]
        proxyId:
          type: ["integer", "null"]
        retailer:
          type: string
        productUrl:
          type: string
        productKeywords:
          type: string
        size:
          type: string
        quantity:
          type: integer
        monitorDelay:
          type: integer
        retryCount:
          type: integer
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - retailer
        - productUrl
        - productKeywords
        - size
        - quantity
        - monitorDelay
        - retryCount
        - status
        - createdAt
        - updatedAt
    CreateTaskBody:
      type: object
      properties:
        groupId:
          type: integer
        profileId:
          type: integer
        proxyId:
          type: integer
        retailer:
          type: string
        productUrl:
          type: string
        productKeywords:
          type: string
        size:
          type: string
        quantity:
          type: integer
        monitorDelay:
          type: integer
        retryCount:
          type: integer
      required:
        - retailer
    UpdateTaskBody:
      type: object
      properties:
        groupId:
          type: integer
        profileId:
          type: integer
        proxyId:
          type: integer
        retailer:
          type: string
        productUrl:
          type: string
        productKeywords:
          type: string
        size:
          type: string
        quantity:
          type: integer
        monitorDelay:
          type: integer
        retryCount:
          type: integer
        status:
          type: string
...